webapp/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
toml==0.10.2
bcrypt==3.2.0
pyjwt